adds scheduled-checkouts codes
