+++ src/controllers/bookingsController.js
@@ -1,6 +1,7 @@
 import models from '../models';
 import getUserData from '../helper/tokenToData';
 import notFound from '../utils/Errors/notFoundRequestError';
+import {Op} from 'sequelize';
 
 const showBookings = async (req, res, next) => {
   try {
@@ -14,8 +15,50 @@ const showBookings = async (req, res, next) => {
     next(error);
   }
 };
+const expiredBookings = async () => {
+  try {
+    const booking = await models.Booking.findAll({
+        where: {
+        To: {
+          [Op.lte]: new Date()
+        },
+        checkedout: false
+      }
+    });
+    if (!booking[0]) {
+      throw new notFound('You do not have any bookings');
+    }
+    booking.forEach(async singleBooking => {
+      // 
+      await models.Accommodation.increment(
+        {
+          numberOfRooms: +1
+        }, {
+          where: {
+          id: singleBooking.accommodationId
+        }
+      })
+      // 
+    });
+    await models.Booking.update(
+      {
+        checkedout: true
+      }, {
+        where: {
+        To: {
+          [Op.lte]: new Date()
+        },
+        checkedout: false
+      }
+    })
+  } catch (error) {
+    console.log(error);
+  }
+};
 
 
-export default showBookings ;
-
+export {
+  showBookings,
+  expiredBookings
+};
